--- listing_gui/__init__.py
@@ -1,3 +1,0 @@
-from .app import main as app_main
-
-__all__ = ["app_main"]
