+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "notion-operator"
 version = "0.1.0"
-description = "Notion連携 出品・ラベル印刷ツール"
+description = "Notion連携 ラベル印刷・データ出力ツール"
 requires-python = ">=3.10"
 dependencies = [
     "flet[all]==0.28.3",
@@ -18,8 +18,6 @@ dependencies = [
     "openpyxl>=3.1.2",
     "python-dotenv>=1.0.0",
     "pydantic>=2.0.0",
-    "selenium>=4.18.0",
-    "webdriver-manager>=4.0.0",
     "brother>=2.3.0",
     "reportlab>=4.0.0",
 ]
