+++ requirements.txt
@@ -13,7 +13,5 @@ pandas>=2.1.0
 openpyxl>=3.1.2
 python-dotenv>=1.0.0
 pydantic>=2.0.0
-selenium>=4.18.0
-webdriver-manager>=4.0.0
 brother>=2.3.0
 reportlab>=4.0.0
